Refactor: ILoginPayload 인터페이스 위치 변경 및 LoginForm 수정
