+++ src/app/login/page.tsx
@@ -1,4 +1,4 @@
-import LoginForm from "./login/LoginForm";
+import LoginForm from "./panel/LoginForm";
 
 export default function LoginPage() {
   return (
