+++ src/types/ILoginPayload.ts
@@ -1,0 +1,4 @@
+interface ILoginPayload {
+  id: string;
+  password: string;
+}
